Implement global config file

+++ hardware/sfp_driver.py
@@ -144,7 +144,7 @@ class Sfp():
 
         diagnostics_block = self.i2c_bus.read_i2c_block_data(SFP_I2C_DIAG_ADDRESS, SFP_DIAG_REG_START, DIAG_DATA_LENGTH)
         temp_bytes = diagnostics_block[TEMP_OFFSET: VCC_OFFSET]
-        temp_h = np.int8(temp_bytes[0])  # signed int8
+        temp_h = float(np.int8(temp_bytes[0]))  # signed int8
         temp_l = self.convert_to_fp(temp_bytes[1], 8, 256)
         temp = temp_h + temp_l
         # print(f"Temp: {temp}")
@@ -156,7 +156,7 @@ class Sfp():
 
         # next is tx output power in mW
         tx_bytes = diagnostics_block[TX_POWER_OFFSET:RX_POWER_OFFSET]
-        tx_power = np.uint16((tx_bytes[0] << 8) | tx_bytes[1]) / 10000  # LSB is equal to 100 uW, so the range is [0, 6.5535]mW
+        tx_power = float(np.uint16((tx_bytes[0] << 8) | tx_bytes[1]) / 10000)  # LSB is equal to 100 uW, so the range is [0, 6.5535]mW
         print(f"Tx power: {tx_power}mW")
         self.data["diagnostics"]["tx_power"] = tx_power
         tx_power_dbm = round(self.convert_to_dB(tx_power), 3)
@@ -165,7 +165,7 @@ class Sfp():
 
         # next is rx received power in mW
         rx_bytes = diagnostics_block[RX_POWER_OFFSET:]
-        rx_power = np.uint16((rx_bytes[0] << 8) | rx_bytes[1]) / 10000 # LSB is equal to 100 uW, so the range is [0, 6.5535]mW
+        rx_power = float(np.uint16((rx_bytes[0] << 8) | rx_bytes[1]) / 10000) # LSB is equal to 100 uW, so the range is [0, 6.5535]mW
         print(f"Rx power: {rx_power}mW")
         self.data["diagnostics"]["rx_power"] = rx_power
         rx_power_dbm = round(self.convert_to_dB(rx_power), 3)
